Use TEXINPUTS rather than relying on relative paths
The fact that \RequirePackage{} takes a path is just because of a lack
of stringent error checking:
https://tex.stackexchange.com/a/124013
  Therefore, this commit switches to the safer method of adding Classes/
and StyleFiles/ to the search path. This is a bit of a pain, in that the
environment variable needs to be set before compilation, but saves a lot
of hassle elsewhere - in particular, we can now import style files into
classes or one another.