Switch to adjustbox package
Seems much safer than the \Gin@nat@width and \maxwidth approach. The
problem with \Gin@nat@width is that it doesn't seem possible to use it
to give relative adjustments to image sizes, which is a major pain. The
adjustbox package makes those operations easier and doesn't expose so
much of the plumbing!