Restructured colour definitions
These are now all defined in lt_basics.sty, when xcolor is loaded. Then
used as necessary by the other packages (or could be overwritten if
desired). Makes it a bit easier to find (i.e. not spread throughout
lt_hyperref.sty and lt_article.cls) and got the profusion of gray /
lightgray under control...
  One potential issues is that the print option to lt_hyperref doesn't
change the AccentColour, but that is splitting hairs.