Add `lt_minimal.sty`
Very lightweight package that is an *almost* drop-in replacement for the
core packages/commands defined in `lt_basics.sty`, `lt_floats.sty`, and
`lt_hyperref.sty`. Coupled with the simplifications in `lt_maths.sty`,
this means it is possible to compile a very lightweight alternative to
the fully formatted output.